requirements.txt
annotated-types==0.6.0
anyio==4.2.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dishka==1.1.1
distro==1.9.0
fastapi==0.109.2
google-auth==2.27.0
google-auth-oauthlib==1.2.0
gspread==6.0.2
gspread-asyncio==2.0.0
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
jinja2==3.1.4
loguru==0.7.2
Markdown==3.6
markdown2==2.4.13
MarkupSafe==2.1.5
oauthlib==3.2.2
openai==1.12.0
packaging==24.0
pluggy==1.5.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.1
pydantic-core==2.16.2
pytest==8.1.1
python-multipart==0.0.9
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sniffio==1.3.0
starlette==0.36.3
StrEnum==0.4.15
tqdm==4.66.2
typing-extensions==4.9.0
urllib3==2.2.0
uvicorn==0.27.1
